Project Initialisation (#1)
* Add issue templates

Create some templates for issues

* Revise README for Card Battle Game details

Updated project title and added detailed game rules and commands for the Card Battle Game.

Signed-off-by: Ischim <marc.ischi@hes-so.ch>

* Update README.md

Move the entry [Tables of contenents] to the first place in Tables of contenets

Signed-off-by: Ischim <marc.ischi@hes-so.ch>

* Update .gitignore

* Add and modify issue template

* Modification of README.md with footer and add a picture SUS X Club Penguin Club

* Modification of README.md

* Reorganisation of icons in README.md

* Revise bug report template for clarity

Updated the bug report template to include application information and removed smartphone details because this pratical work is not destined to use on smartphone.

Signed-off-by: Ischim <marc.ischi@hes-so.ch>

* Update README with AI tools utilization

Added details about AI tools and their usage in the README.

Signed-off-by: Ischim <marc.ischi@hes-so.ch>

* Update README to remove unnecessary text

Removed an @ becouse the description is an copy of our discussion in DAI Course

Signed-off-by: Ischim <marc.ischi@hes-so.ch>

---------

Signed-off-by: Ischim <marc.ischi@hes-so.ch>